introduce component to show list of cell style options
